mini-projet-users/backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg2-binary

python-dotenv

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose[cryptography]
email-validator
python-multipart
